requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
Pillow>=8.0.0
PyYAML>=5.4.0
tensorboard>=2.4.0
